+++ CHANGES.rst
@@ -33,6 +33,16 @@ mast
 - Expand the supported data types for filter values in ``Mast.mast_query``. Previously, users had to input
   filter values enclosed in lists, even when specifying a single value or dictionary. [#3422]
 
+jplspec
+^^^^^^^
+
+- Moved to linelists/.  astroquery.jplspec is now deprecated in favor of astroquery.linelists.jplspec [#3455]
+
+linelists.jplspec
+^^^^^^^^^^^^^^^^^
+
+- New location for jplspec.  astroquery.jplspec is now deprecated in favor of astroquery.linelists.jplspec [#3455]
+
 
 Infrastructure, Utility and Other Changes and Additions
 -------------------------------------------------------